feat(assignment-02): added simple calculator in control flow

+++ Assignment-01_python_basic_concept/math_operation.py
@@ -12,7 +12,7 @@ try:
 
     # Handle division safely
     if num2 != 0:
-        division = num1 / num2
+        division = float(num1 / num2)
     else:
         division = "Undefined (Cannot divide by zero)"
